refactor(mails): split mails.test.ts into imap.test.ts + http.test.ts
Mirrors the source split from #700 — tests co-locate with the file
they cover. 14 describe blocks re-partitioned:

  imap.test.ts (9)  — STORE ops, buildFlagSetClause, setMailFlags,
                      expungeDeletedMails, buildCriterionClause,
                      searchMailsByUid
  http.test.ts (5)  — getAccountStats, draft-exclusion invariant,
                      buildHeaderAddressCondition, getMailHeaders,
                      getMailHeadersDelta

Two mechanical rewrites required after the move:

- `await import("./mails")` → `await import(".")` since the tests now
  live INSIDE `mails/` (dir index).
- `path.join(import.meta.dir, "mails")` → `import.meta.dir` for the
  source-scanning tests, since the test file's directory now IS the
  mails source dir. Added a `.test.ts` filter to the readdir so the
  test file itself doesn't get concatenated into the source-under-test
  and produce false positives on regex patterns quoted in the test.

All 77 repository tests pass; full server suite 1217/0.
