+++ src/server/lib/postgres/repositories/mails/imap.test.ts
@@ -172,7 +172,7 @@ describe("buildFlagSetClause — empty/unknown-only STORE is a no-op (#671)", ()
   // collided with the trailing `updated = CURRENT_TIMESTAMP` → Postgres
   // "multiple assignments to same column". setMailFlags treats "" as the
   // legal no-op path (RFC 3501 §6.4.6).
-  const load = async () => (await import("./mails")).buildFlagSetClause;
+  const load = async () => (await import(".")).buildFlagSetClause;
 
   it("returns '' for +FLAGS with an empty flag list", async () => {
     const buildFlagSetClause = await load();
@@ -228,9 +228,9 @@ describe("setMailFlags — no-op STORE skips the UPDATE (source regression for #
     const path = await import("path");
     const mailsSource = (
       await Promise.all(
-        (await fs.readdir(path.join(import.meta.dir, "mails"))).sort()
-          .filter((f) => f.endsWith(".ts"))
-          .map((f) => fs.readFile(path.join(import.meta.dir, "mails", f), "utf8"))
+        (await fs.readdir(import.meta.dir)).sort()
+          .filter((f) => f.endsWith(".ts") && !f.endsWith(".test.ts"))
+          .map((f) => fs.readFile(path.join(import.meta.dir, f), "utf8"))
       )
     ).join("\n");
     const fnMatch = mailsSource.match(/export const setMailFlags[\s\S]*?\n};/);
@@ -270,9 +270,9 @@ describe("expungeDeletedMails — `updated` column refresh (regression for #456)
     const path = await import("path");
     mailsSource = (
       await Promise.all(
-        (await fs.readdir(path.join(import.meta.dir, "mails"))).sort()
-          .filter((f) => f.endsWith(".ts"))
-          .map((f) => fs.readFile(path.join(import.meta.dir, "mails", f), "utf8"))
+        (await fs.readdir(import.meta.dir)).sort()
+          .filter((f) => f.endsWith(".ts") && !f.endsWith(".test.ts"))
+          .map((f) => fs.readFile(path.join(import.meta.dir, f), "utf8"))
       )
     ).join("\n");
     const fnMatch = mailsSource.match(
@@ -308,176 +308,13 @@ describe("expungeDeletedMails — `updated` column refresh (regression for #456)
   });
 });
 
-describe("getAccountStats — envelope_to inclusion in received address expansion", () => {
-  // Mails sent via listserv-style routing (e.g. GitHub notifications) carry
-  // a MIME `to` header that points at the list address (e.g.
-  // `<budget@noreply.github.com>`) and an SMTP-level `envelope_to` that
-  // points at the actual recipient sub-address (e.g. `<x@hoie.kim>`). If
-  // the received-side address expansion ignores envelope_to, those mails
-  // never surface in the per-account view — but the push badge counts
-  // them via the broader `getUnreadNotifications` query, so the FE shows
-  // 0 unread while the iOS badge shows N. Verified against prod on
-  // 2026-05-23: admin had badge=26 / FE=0 because 26 GitHub notification
-  // mails carried envelope_to=claoie@hoie.kim but MIME to=noreply.github.
-  let mailsSource: string;
-  let fnSource: string;
-
-  beforeAll(async () => {
-    const fs = await import("fs/promises");
-    const path = await import("path");
-    mailsSource = (
-      await Promise.all(
-        (await fs.readdir(path.join(import.meta.dir, "mails"))).sort()
-          .filter((f) => f.endsWith(".ts"))
-          .map((f) => fs.readFile(path.join(import.meta.dir, "mails", f), "utf8"))
-      )
-    ).join("\n");
-    const fnMatch = mailsSource.match(
-      /export const getAccountStats[\s\S]*?\n};/
-    );
-    if (!fnMatch) throw new Error("getAccountStats not found in mails/*.ts");
-    fnSource = fnMatch[0];
-  });
-
-  it("received-branch address expansion unions envelope_to with to/cc/bcc", () => {
-    // The received-branch expansion lives in the shared
-    // RECEIVED_ADDRESS_EXPANSION constant (reused by getAccountStats and
-    // searchAccountStats so the two never drift). getAccountStats' received
-    // branch must reference it.
-    const constMatch = mailsSource.match(
-      /const\s+RECEIVED_ADDRESS_EXPANSION\s*=\s*`([^`]*)`/
-    );
-    if (!constMatch) throw new Error("RECEIVED_ADDRESS_EXPANSION not found");
-    const receivedSql = constMatch[1];
-    expect(receivedSql).toContain("to_address");
-    expect(receivedSql).toContain("cc_address");
-    expect(receivedSql).toContain("bcc_address");
-    expect(receivedSql).toContain("envelope_to");
-    expect(fnSource).toContain("RECEIVED_ADDRESS_EXPANSION");
-  });
-
-  it("received-branch null-check includes envelope_to", () => {
-    // Otherwise rows with only envelope_to populated (no MIME recipient
-    // headers, which happens for some listserv-style senders) would be
-    // filtered out before the address expansion even fires. Also a shared
-    // constant so searchAccountStats inherits the same null-check.
-    const constMatch = mailsSource.match(
-      /const\s+RECEIVED_ADDRESS_NOT_NULL\s*=\s*`([^`]*)`/
-    );
-    if (!constMatch) throw new Error("RECEIVED_ADDRESS_NOT_NULL not found");
-    const receivedSql = constMatch[1];
-    expect(receivedSql).toContain("to_address IS NOT NULL");
-    expect(receivedSql).toContain("envelope_to IS NOT NULL");
-    expect(fnSource).toContain("RECEIVED_ADDRESS_NOT_NULL");
-  });
-
-  it("searchAccountStats reuses the shared received-address constants + full-text predicate", () => {
-    // The search side-tab must list exactly the accounts whose mail appears in
-    // the search results, so searchAccountStats' account attribution has to
-    // match getAccountStats' received path (same envelope_to union) AND filter
-    // to the search term via the same tsquery searchMails uses.
-    const fnMatch = mailsSource.match(
-      /export const searchAccountStats[\s\S]*?\n};/
-    );
-    if (!fnMatch) throw new Error("searchAccountStats not found in mails/*.ts");
-    const src = fnMatch[0];
-    expect(src).toContain("RECEIVED_ADDRESS_EXPANSION");
-    expect(src).toContain("RECEIVED_ADDRESS_NOT_NULL");
-    expect(src).toContain("search_vector @@ plainto_tsquery('english', $2)");
-    expect(src).toContain("expunged = FALSE");
-    expect(src).toContain("draft = FALSE");
-  });
-
-  it("sent-branch address expansion remains from_address only", () => {
-    // Don't accidentally widen the sent view — envelope_from has its own
-    // semantics (bounce path) and isn't symmetric with envelope_to here.
-    const exprMatch = fnSource.match(
-      /const\s+addressExpansion\s*=\s*useSentExpansion\s*\?\s*`([^`]*)`/
-    );
-    if (!exprMatch) throw new Error("sent branch not found");
-    const sentSql = exprMatch[1];
-    expect(sentSql).toContain("from_address");
-    expect(sentSql).not.toContain("envelope_to");
-    expect(sentSql).not.toContain("envelope_from");
-  });
-
-  it("spamOnly forces the received expansion and filters is_spam", () => {
-    // Spam is received mail grouped per receiving account, so spamOnly must
-    // never take the sent (from_address) expansion, and must restrict to
-    // is_spam received rows so the per-account spam counts match the folder.
-    expect(fnSource).toContain("const useSentExpansion = sent && !spamOnly;");
-    const spamMatch = fnSource.match(
-      /const\s+spamCondition\s*=\s*spamOnly\s*\?\s*`([^`]*)`/
-    );
-    if (!spamMatch) throw new Error("spamCondition not found");
-    expect(spamMatch[1]).toContain("is_spam = TRUE");
-    expect(spamMatch[1]).toContain("sent = FALSE");
-  });
-});
-
-describe("draft-exclusion invariant — user-facing read paths hide drafts (#611)", () => {
-  // A draft lives in the IMAP Drafts folder; the web client presents no Drafts
-  // view, so a draft must not surface in ANY user-facing read surface (folder
-  // lists, per-account counts, search results, push badge, spam list). The
-  // invariant is enforced query-side with `AND draft = FALSE`. getMailHeaders
-  // and getAccountStats already carried it; searchMails / getUnreadNotifications
-  // dropped it (#611) so a draft showed in search but in no folder/count. The
-  // spam list is now a per-account getMailHeaders query (`?spam=1`), so its
-  // draft filter is covered by the getMailHeaders guard below. These source-scan
-  // guards pin the filter on every path so they cannot re-drift independently.
-  // Source-text scanning (not a live query) is
-  // used here because module-mock interactions make pool.query mocking fragile
-  // in the full suite — same rationale as the getAccountStats guard above.
-  let mailsSource: string;
-
-  beforeAll(async () => {
-    const fs = await import("fs/promises");
-    const path = await import("path");
-    mailsSource = (
-      await Promise.all(
-        (await fs.readdir(path.join(import.meta.dir, "mails"))).sort()
-          .filter((f) => f.endsWith(".ts"))
-          .map((f) => fs.readFile(path.join(import.meta.dir, "mails", f), "utf8"))
-      )
-    ).join("\n");
-  });
-
-  const bodyOf = (name: string): string => {
-    const re = new RegExp(`export const ${name}[\\s\\S]*?\\n};`);
-    const m = mailsSource.match(re);
-    if (!m) throw new Error(`${name} not found in mails/*.ts`);
-    return m[0];
-  };
-
-  // Every user-facing read path that must hide drafts.
-  const draftHidingPaths = [
-    "getMailHeaders",
-    "getAccountStats",
-    "searchMails",
-    "getUnreadNotifications",
-  ];
-
-  for (const name of draftHidingPaths) {
-    it(`${name} filters draft = FALSE`, () => {
-      expect(bodyOf(name)).toMatch(/draft = FALSE/);
-    });
-  }
-
-  it("searchMails keeps the draft filter alongside its existing expunged filter", () => {
-    // Guard against a regression that removes one filter while editing the other.
-    const body = bodyOf("searchMails");
-    expect(body).toMatch(/expunged = FALSE/);
-    expect(body).toMatch(/draft = FALSE/);
-  });
-});
-
 describe("buildCriterionClause — flag criteria use schema columns", () => {
   // Regression guard (originally on searchMailsByUid's inline switch, retargeted
   // here when #551 extracted the per-criterion logic into buildCriterionClause):
   // the answered/deleted/draft flags map to their real boolean columns, never to
   // a bare "FALSE" match-none sentinel that an earlier draft of this fix used.
   const clauseFor = async (type: string) => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     return buildCriterionClause({ type }, "uid_account", values as never);
   };
@@ -520,131 +357,6 @@ describe("buildCriterionClause — flag criteria use schema columns", () => {
   });
 });
 
-describe("buildHeaderAddressCondition — envelope_to in received-branch address condition", () => {
-  // Mails addressed via envelope_to (e.g. GitHub notification routing,
-  // listserv sub-addressing) must appear in per-account mail lists, not
-  // only in account-stats counts. The received-branch filter must include
-  // envelope_to alongside MIME to/cc/bcc. The condition is shared by the
-  // full-list (getMailHeaders) and delta (getMailHeadersDelta) paths via the
-  // extracted buildHeaderAddressCondition helper, so this guards the one
-  // source of truth.
-  let mailsSource: string;
-  let fnSource: string;
-
-  beforeAll(async () => {
-    const fs = await import("fs/promises");
-    const path = await import("path");
-    mailsSource = (
-      await Promise.all(
-        (await fs.readdir(path.join(import.meta.dir, "mails"))).sort()
-          .filter((f) => f.endsWith(".ts"))
-          .map((f) => fs.readFile(path.join(import.meta.dir, "mails", f), "utf8"))
-      )
-    ).join("\n");
-    const fnMatch = mailsSource.match(
-      /export const buildHeaderAddressCondition[\s\S]*?\n};/
-    );
-    if (!fnMatch) throw new Error("buildHeaderAddressCondition not found in mails/*.ts");
-    fnSource = fnMatch[0];
-  });
-
-  it("received branch's addressCondition unions envelope_to with to/cc/bcc", () => {
-    // Source uses `${TO_ADDRESS}` template-literal substitution that
-    // expands to "to_address" at runtime; the static text contains the
-    // token, so the test asserts on the template tokens directly.
-    const exprMatch = fnSource.match(
-      /receivedCondition\s*=\s*`([^`]*)`/
-    );
-    if (!exprMatch) throw new Error("receivedCondition not found");
-    const receivedSql = exprMatch[1];
-    expect(receivedSql).toContain("${TO_ADDRESS}");
-    expect(receivedSql).toContain("cc_address @>");
-    expect(receivedSql).toContain("bcc_address @>");
-    expect(receivedSql).toContain("envelope_to @>");
-  });
-
-  it("sent branch's addressCondition remains from_address only", () => {
-    const exprMatch = fnSource.match(/sentCondition\s*=\s*`([^`]*)`/);
-    if (!exprMatch) throw new Error("sentCondition not found");
-    const sentSql = exprMatch[1];
-    expect(sentSql).toContain("${FROM_ADDRESS}");
-    expect(sentSql).not.toContain("envelope_to");
-    expect(sentSql).not.toContain("envelope_from");
-  });
-});
-
-describe("getMailHeaders / getMailHeadersDelta — `?since=` delta path (#457)", () => {
-  // The delta path drives the IndexedDB cache: only rows changed since the
-  // client's cursor, plus tombstones for rows expunged in that window. These
-  // guard the SQL invariants that make the cursor safe — they can't be unit-
-  // checked without a DB, so we assert on the query source (the repo's
-  // established style for SQL-shape regressions).
-  let mailsSource: string;
-
-  beforeAll(async () => {
-    const fs = await import("fs/promises");
-    const path = await import("path");
-    mailsSource = (
-      await Promise.all(
-        (await fs.readdir(path.join(import.meta.dir, "mails"))).sort()
-          .filter((f) => f.endsWith(".ts"))
-          .map((f) => fs.readFile(path.join(import.meta.dir, "mails", f), "utf8"))
-      )
-    ).join("\n");
-  });
-
-  it("getMailHeaders filters on `updated >` when options.since is set", () => {
-    const fnMatch = mailsSource.match(/export const getMailHeaders[\s\S]*?\n};/);
-    if (!fnMatch) throw new Error("getMailHeaders not found");
-    const src = fnMatch[0];
-    // Guarded behind options.since so the full-list path is unaffected.
-    expect(src).toContain("options.since !== undefined");
-    expect(src).toContain("AND updated > $");
-  });
-
-  it("getMailHeadersDelta scans expunged rows for tombstones over the same window", () => {
-    const fnMatch = mailsSource.match(
-      /export const getMailHeadersDelta[\s\S]*?\n};/
-    );
-    if (!fnMatch) throw new Error("getMailHeadersDelta not found");
-    const src = fnMatch[0];
-    // Tombstone query: expunged rows in the same account, changed in-window.
-    expect(src).toContain("expunged = TRUE");
-    expect(src).toContain("AND updated > $3");
-    // Reuses the one address-condition source of truth, not a private copy.
-    expect(src).toContain("buildHeaderAddressCondition(options)");
-  });
-
-  it("getMailHeadersDelta reads as_of from the DB clock (with safety margin) BEFORE the data queries", () => {
-    const fnMatch = mailsSource.match(
-      /export const getMailHeadersDelta[\s\S]*?\n};/
-    );
-    if (!fnMatch) throw new Error("getMailHeadersDelta not found");
-    const src = fnMatch[0];
-    // as_of must come from now() (same timeline as the `updated` column set by
-    // CURRENT_TIMESTAMP), not the app clock, or clock skew could skip rows.
-    const asOfIdx = src.indexOf("now()");
-    const headersIdx = src.indexOf("getMailHeaders(");
-    expect(asOfIdx).toBeGreaterThanOrEqual(0);
-    expect(headersIdx).toBeGreaterThanOrEqual(0);
-    // Captured first → a safe lower bound (at-least-once on concurrent writes).
-    expect(asOfIdx).toBeLessThan(headersIdx);
-    // Backed off by a safety margin so the commit-latency / skew window re-sends
-    // rather than skips.
-    expect(src).toContain("make_interval(secs => $1)");
-    expect(src).toContain("DELTA_CURSOR_SAFETY_MARGIN_SECONDS");
-  });
-
-  it("getMailHeadersDelta echoes `since` as as_of on failure (cursor must not advance)", () => {
-    const fnMatch = mailsSource.match(
-      /export const getMailHeadersDelta[\s\S]*?\n};/
-    );
-    if (!fnMatch) throw new Error("getMailHeadersDelta not found");
-    const src = fnMatch[0];
-    expect(src).toMatch(/return\s*\{\s*as_of:\s*since/);
-  });
-});
-
 describe("buildCriterionClause — NOT/OR SQL generation (regression for #551)", () => {
   // buildCriterionClause receives the normalised `{ type, value }` shape that
   // store.ts's simplifyCriterion produces. It pushes bound params onto `values`
@@ -652,7 +364,7 @@ describe("buildCriterionClause — NOT/OR SQL generation (regression for #551)",
   // or null when the criterion imposes no constraint.
 
   it("NOT wraps the inner clause instead of dropping it", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     const frag = buildCriterionClause(
       { type: "NOT", value: { type: "SEEN" } },
@@ -666,7 +378,7 @@ describe("buildCriterionClause — NOT/OR SQL generation (regression for #551)",
   });
 
   it("OR joins both sides with continuous param numbering", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     const frag = buildCriterionClause(
       {
@@ -684,7 +396,7 @@ describe("buildCriterionClause — NOT/OR SQL generation (regression for #551)",
   });
 
   it("NOT FROM negates a text predicate and binds its param", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     const frag = buildCriterionClause(
       { type: "NOT", value: { type: "FROM", value: "spam@x" } },
@@ -696,7 +408,7 @@ describe("buildCriterionClause — NOT/OR SQL generation (regression for #551)",
   });
 
   it("continues param numbering from an already-populated values array", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = ["user-1", false]; // e.g. base user_id/sent params
     const frag = buildCriterionClause(
       {
@@ -714,7 +426,7 @@ describe("buildCriterionClause — NOT/OR SQL generation (regression for #551)",
   });
 
   it("drops an OR whose side imposes no constraint rather than over-narrowing", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     const frag = buildCriterionClause(
       {
@@ -732,7 +444,7 @@ describe("buildCriterionClause — NOT/OR SQL generation (regression for #551)",
   });
 
   it("normalised NOT BEFORE flows a Date param through correctly", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     const when = new Date("2026-01-01T00:00:00Z");
     const frag = buildCriterionClause(
@@ -745,7 +457,7 @@ describe("buildCriterionClause — NOT/OR SQL generation (regression for #551)",
   });
 
   it("plain criteria are unaffected by the refactor", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     expect(buildCriterionClause({ type: "SEEN" }, "uid_account", values as never)).toBe(
       "read = TRUE"
@@ -761,7 +473,7 @@ describe("buildCriterionClause — UID_SET ORs its ranges (#659)", () => {
   // always-empty set. Now the whole set renders as a single OR-of-ranges.
 
   it("renders a single exact element without an OR wrapper", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     const frag = buildCriterionClause(
       { type: "UID_SET", value: [{ start: 5 }] },
@@ -773,7 +485,7 @@ describe("buildCriterionClause — UID_SET ORs its ranges (#659)", () => {
   });
 
   it("ORs disjoint exact elements (`1,3`) instead of ANDing to empty", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     const frag = buildCriterionClause(
       { type: "UID_SET", value: [{ start: 1 }, { start: 3 }] },
@@ -785,7 +497,7 @@ describe("buildCriterionClause — UID_SET ORs its ranges (#659)", () => {
   });
 
   it("ORs mixed exact + range elements (`2:3,5:7`) with parenthesised ranges", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     const frag = buildCriterionClause(
       { type: "UID_SET", value: [{ start: 2, end: 3 }, { start: 5, end: 7 }] },
@@ -799,7 +511,7 @@ describe("buildCriterionClause — UID_SET ORs its ranges (#659)", () => {
   });
 
   it("renders a lone range without an OR wrapper (common `1:3` client form)", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     const frag = buildCriterionClause(
       { type: "UID_SET", value: [{ start: 1, end: 3 }] },
@@ -811,7 +523,7 @@ describe("buildCriterionClause — UID_SET ORs its ranges (#659)", () => {
   });
 
   it("ANDs correctly against a sibling flag key (`SEEN 1,3`)", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     // Sibling keys are joined with AND by searchMailsByUid; the set stays a
     // single OR-group so the intersection is `read AND (uid∈{1,3})`.
     const values: unknown[] = [];
@@ -828,7 +540,7 @@ describe("buildCriterionClause — UID_SET ORs its ranges (#659)", () => {
   });
 
   it("imposes no constraint for an empty set (caller skips it)", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     expect(
       buildCriterionClause({ type: "UID_SET", value: [] }, "uid_account", values as never)
@@ -850,9 +562,9 @@ describe("searchMailsByUid — no result cap (#553)", () => {
     const path = await import("path");
     const mailsSource = (
       await Promise.all(
-        (await fs.readdir(path.join(import.meta.dir, "mails"))).sort()
-          .filter((f) => f.endsWith(".ts"))
-          .map((f) => fs.readFile(path.join(import.meta.dir, "mails", f), "utf8"))
+        (await fs.readdir(import.meta.dir)).sort()
+          .filter((f) => f.endsWith(".ts") && !f.endsWith(".test.ts"))
+          .map((f) => fs.readFile(path.join(import.meta.dir, f), "utf8"))
       )
     ).join("\n");
     const fnMatch = mailsSource.match(
@@ -878,7 +590,7 @@ describe("buildCriterionClause — BODY/TEXT search the message body (#552)", ()
   // (plain-text body) column and missed virtually every body-content match.
 
   it("BODY matches the body column only", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     const frag = buildCriterionClause(
       { type: "BODY", value: "needle" },
@@ -893,7 +605,7 @@ describe("buildCriterionClause — BODY/TEXT search the message body (#552)", ()
   });
 
   it("TEXT matches header columns plus the body column", async () => {
-    const { buildCriterionClause } = await import("./mails");
+    const { buildCriterionClause } = await import(".");
     const values: unknown[] = [];
     const frag = buildCriterionClause(
       { type: "TEXT", value: "needle" },
@@ -904,54 +616,5 @@ describe("buildCriterionClause — BODY/TEXT search the message body (#552)", ()
     expect(frag).toContain("from_text ILIKE");
     expect(frag).toContain("to_text ILIKE");
     expect(frag).toContain("text ILIKE");
-  });
-});
-
-describe("getMailHeaders — saved query spans both folders (#568)", () => {
-  // A starred mail can be either sent or received. A saved query with no
-  // explicit folder must match an account address in EITHER from_address
-  // (sent) or the received to/cc/bcc/envelope_to branch — otherwise a
-  // starred sent mail is unreachable from the Saved view, the client-side
-  // complement of #384's server fix.
-  let fnSource: string;
-
-  beforeAll(async () => {
-    const fs = await import("fs/promises");
-    const path = await import("path");
-    const mailsSource = (
-      await Promise.all(
-        (await fs.readdir(path.join(import.meta.dir, "mails"))).sort()
-          .filter((f) => f.endsWith(".ts"))
-          .map((f) => fs.readFile(path.join(import.meta.dir, "mails", f), "utf8"))
-      )
-    ).join("\n");
-    const fnMatch = mailsSource.match(
-      /export const buildHeaderAddressCondition[\s\S]*?\n};/
-    );
-    if (!fnMatch) throw new Error("buildHeaderAddressCondition not found in mails/*.ts");
-    fnSource = fnMatch[0];
-  });
-
-  it("uses the union (sent OR received) condition when saved && !sent", () => {
-    const exprMatch = fnSource.match(
-      /return\s+([\s\S]*?);/
-    );
-    if (!exprMatch) throw new Error("address-condition return expression not found");
-    const expr = exprMatch[1];
-    // The saved-and-not-sent branch is the union of both folder conditions.
-    expect(expr).toContain("options.saved && !options.sent");
-    expect(expr).toContain("sentCondition} OR ${receivedCondition");
-  });
-
-  it("falls back to the sent-only or received-only condition otherwise", () => {
-    const exprMatch = fnSource.match(
-      /return\s+([\s\S]*?);/
-    );
-    if (!exprMatch) throw new Error("address-condition return expression not found");
-    const expr = exprMatch[1];
-    expect(expr).toContain("options.sent");
-    // Non-union branches reuse the single-folder conditions verbatim.
-    expect(expr).toMatch(/\?\s*sentCondition/);
-    expect(expr).toContain(": receivedCondition");
   });
 });
